requirements.txt
setuptools==66.0.0
apache-airflow==2.6.3
apache-airflow[postgres]==2.6.3
apache-airflow[cncf.kubernetes]==2.6.3
apache-airflow[virtualenv]==2.6.3
psycopg2-binary==2.9.9
SQLAlchemy==1.4.50
kafka-python==2.0.2
pendulum==2.1.2
connexion[swagger-ui]==2.14.2
openapi-spec-validator<0.6.0
Flask-Session==0.5.0
requests==2.31.0
numpy==1.23.5
pandas==1.5.3
pyspark==3.3.2
cassandra-driver==3.28.0
Flask-SQLAlchemy==2.5.1
flask>=2.2.5,<2.3.0
flask-login>=0.6.2
flask-wtf>=0.15.1
wtforms>=2.3.3
alembic<2.0.0
python-dateutil>=2.8.2
werkzeug>=2.2.3,<2.3.0
markupsafe>=2.1.1
kubernetes>=10.0.1
virtualenv>=20.0.0
Flask-Limiter[redis]==3.3.0
redis==4.5.5
